+++ src/Feature/Sms/Bag/ScheduleSmsBag.php
@@ -64,6 +64,18 @@ class ScheduleSmsBag
         return $this;
     }
 
+    public function setExternalId(string $idx, bool $checkIdx = null): self
+    {
+        $this->idx = [$idx];
+        $this->checkIdx = $checkIdx;
+
+        return $this;
+    }
+
+    /**
+     * @deprecated
+     * @see ScheduleSmsBag::setExternalId()
+     */
     public function setIdx(array $idx, bool $checkIdx = null): self
     {
         $this->idx = $idx;