+++ src/Feature/Sms/Bag/ScheduleSmsToGroupBag.php
@@ -53,6 +53,18 @@ class ScheduleSmsToGroupBag
         return $bag;
     }
 
+    public function setExternalId(string $idx, bool $checkIdx = null): self
+    {
+        $this->idx = [$idx];
+        $this->checkIdx = $checkIdx;
+
+        return $this;
+    }
+
+    /**
+     * @deprecated
+     * @see ScheduleSmsToGroupBag::setExternalId()
+     */
     public function setIdx(array $idx, bool $checkIdx = null): self
     {
         $this->idx = $idx;